improvements to condense_list and to sort_and_print_list
condense list will now put all instances of an item on oone line
even if the units are not the same.  It collects like units, then
stacks them on one line.

sort and print will nowe first check the store config file and then
the default for item locations.  This mean new items usually only need to
be added in the default. the stoer config should only redefine what is not
default